pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rein-ai"
dynamic = ["version"]
description = "Declarative YAML workflow orchestrator for multi-agent AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Rein Contributors"},
]
keywords = ["ai", "agents", "workflow", "orchestrator", "multi-agent", "yaml", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "jsonschema>=4.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
all = ["anthropic>=0.40.0", "openai>=1.0.0"]
daemon = ["websockets>=12.0"]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "websockets>=12.0",
    "mcp>=1.0.0",
]

[project.scripts]
rein = "rein.cli:main"
rein-mcp = "rein.mcp_server:main"

[project.urls]
Homepage = "https://github.com/dklymentiev/rein-orchestrator"
Documentation = "https://github.com/dklymentiev/rein-orchestrator#readme"
Repository = "https://github.com/dklymentiev/rein-orchestrator"
Issues = "https://github.com/dklymentiev/rein-orchestrator/issues"

[tool.setuptools.dynamic]
version = {attr = "rein.__version__"}

[tool.setuptools.packages.find]
include = ["rein", "rein.*", "models", "models.*", "schemas"]
exclude = ["rein-v1-archive*", "tests*"]

[tool.setuptools.package-data]
schemas = ["*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=rein --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # line length handled by formatter
